refactor(filid): move collectSubtreeNodes into usageTracker/utils organ
The re-review's own structure_validate flagged the new helper as a
zero-peer-file at the usageTracker fractal root — a violation this
branch introduced minutes earlier. FCA: new files in a fractal root go
into an organ; mirror the sibling testCoverageChecker/strategies pattern.
